pyproject.toml
[project]
name = "opencloud-mcp"
version = "1.0.0"
description = "OpenCloud MCP Server — WebDAV, CalDAV, CardDAV"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0",
    "uvicorn>=0.30",
    "webdav4>=0.10",
    "caldav>=1.4",
    "vobject>=0.9",
    "python-dotenv>=1.0",
    "cryptography",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
